+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Refactor
 
 - `pimprint-core-bundle` v4.5 compatibility - Symfony service architecture
+- Example: Use PimPrint `CommandQueue` Service directly in `CarsDemo\CarSalesLabel\LabelRenderer`
 
 ## [4.4.0] - 2025-06-18